requirements.txt
Django
djangorestframework
django-rql